% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_strums.R
\name{get_all_strums}
\alias{get_all_strums}
\title{Get all strumming patterns}
\usage{
get_all_strums(notetype, n_min, dlm)
}
\arguments{
\item{notetype}{Type of note. Currently supports 4 (1/4), 8 (1/8) and 16 (1/16) notes}

\item{n_min}{Minimal number of beats}

\item{dlm}{Delimiter for printing}
}
\value{
A tibble containing all beats and strums. First column is beats. Second column is the associated strums
}
\description{
Get all strumming patterns
}
\examples{
get_all_strums(8, 4, '|')
get_all_strums(16, 7, '|')
}
