% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_strum.R
\name{print_strum}
\alias{print_strum}
\title{print strumming patterns}
\usage{
print_strum(df)
}
\arguments{
\item{df}{Dataframe of beats and strums}
}
\description{
Take beats & strums dataframe as input and print each beat and its associated strum to screen
}
\examples{
print_strum(beats_strums)

}
